+++ Presentation.MauiMvvm/ViewModels/ShowCustomersViewModel.cs
@@ -21,11 +21,23 @@ public partial class ShowCustomersViewModel : ObservableObject
     private string _statusMessage = string.Empty;
 
     [ObservableProperty]
+    private Color _statusMessageColor = Colors.Black;
+
+    [ObservableProperty]
     private ObservableCollection<CustomerModel> _customerList;
+
+
+    [ObservableProperty]
+    private bool _isSortedAscendingByCustomerId = true;
+
+    [ObservableProperty]
+    private bool _isSortedAscendingByCustomerName = false;
 
     [RelayCommand]
     private async Task LoadCustomers()
     {
+        StatusMessage = string.Empty;
+
         var result = await _customerService.GetAllCustomersAsync();
 
         if (result.Success && result.Data != null)
@@ -39,6 +51,7 @@ public partial class ShowCustomersViewModel : ObservableObject
         else
         {
             StatusMessage = "Couldn't load users.";
+            StatusMessageColor = Colors.Firebrick;
         }
     }
 
@@ -50,18 +63,66 @@ public partial class ShowCustomersViewModel : ObservableObject
             StatusMessage = "Invalid customer.";
             return;
         }
-        var result = await _customerService.DeleteCustomerAsync(customer.CustomerId);
 
-        if (result.Success)
+        bool confirm = await Shell.Current.DisplayAlert("Confirmation", "Please confirm deletion", "Confirm", "Cancel");
+        if (confirm)
         {
-            CustomerList.Remove(customer);
-            StatusMessage = "Customer deleted successfully.";
+            var result = await _customerService.DeleteCustomerAsync(customer.CustomerId);
+
+            if (result.Success)
+            {
+                CustomerList.Remove(customer);
+                StatusMessage = "Customer deleted successfully.";
+                StatusMessageColor = Colors.Black;
+            }
+            else
+            {
+                StatusMessage = "Customer exists in a project and cannot be deleted.";
+                StatusMessageColor = Colors.Firebrick;
+            }
+        }
+    }
+
+    [RelayCommand]
+    private void SortCustomerListByCustomerId()
+    {
+        if (CustomerList == null)
+            return;
+
+        if (IsSortedAscendingByCustomerId)
+        {
+            var sortedList = CustomerList.OrderByDescending(c => c.CustomerId).ToList();
+            CustomerList = new ObservableCollection<CustomerModel>(sortedList);
         }
         else
         {
-            StatusMessage = "Failed to delete customer.";
+            var sortedList = CustomerList.OrderBy(p => p.CustomerId).ToList();
+            CustomerList = new ObservableCollection<CustomerModel>(sortedList);
         }
+
+        IsSortedAscendingByCustomerId = !IsSortedAscendingByCustomerId;
     }
+
+    [RelayCommand]
+    private void SortCustomerListByCustomerName()
+    {
+        if (CustomerList == null)
+            return;
+
+        if (IsSortedAscendingByCustomerName)
+        {
+            var sortedList = CustomerList.OrderByDescending(c => c.CustomerName).ToList();
+            CustomerList = new ObservableCollection<CustomerModel>(sortedList);
+        }
+        else
+        {
+            var sortedList = CustomerList.OrderBy(p => p.CustomerName).ToList();
+            CustomerList = new ObservableCollection<CustomerModel>(sortedList);
+        }
+
+        IsSortedAscendingByCustomerName = !IsSortedAscendingByCustomerName;
+    }
+
 
     [RelayCommand]
     private async Task NavigateToEditCustomer(CustomerModel customer)
